Changed interpolation type Fixed GtkAdjustment config on size-allocation Fixed scaling issue with resizing -- it couldn't scale on half pixels (half pixel in source pixbuf) if nessecary
(Old svn revision: 2281)

+++ src/picture_viewer.h
@@ -45,12 +45,12 @@ typedef struct _RsttoPictureViewer RsttoPictureViewer;
 
 struct _RsttoPictureViewer
 {
-	GtkWidget        parent;
+	GtkWidget         parent;
 	GdkPixbuf        *src_pixbuf;
 	GdkPixbuf        *dst_pixbuf; /* The pixbuf which ends up on screen */
 	GtkAdjustment    *vadjustment;
 	GtkAdjustment    *hadjustment;
-	gint              scale;
+	gdouble           scale;
 };
 
 typedef struct _RsttoPictureViewerClass RsttoPictureViewerClass;
